separate vm part 1
