+++ Cargo.toml
@@ -20,7 +20,8 @@ num = "0.4"
 futures-lite = "1.12.0"
 itertools = "0.10.5"
 tokio = { version = "1.26.0", features = ["sync"] }
+once_cell = "1.16.0"
 
 [dev-dependencies]
 petgraph = "0.6.3"
-once_cell = "1.16.0"
+
